refactor: :art: organized context files
